+++ corpus/issue-2067/src/my_test2.clj
@@ -1,0 +1,14 @@
+(ns my-test2
+  (:require [foo :as foo]))
+
+(defn my-test [& xs]
+  (prn xs))
+
+(defmacro my-test-macro [x]
+  x)
+
+(my-test
+ (with-redefs [inc dec] 1)
+ (foo/with-redefs [inc dec] 1))
+
+(my-test-macro [])